Move event finalize into event sub-module
